feat(release): new major version and rewritten from scratch.
This is a new version of the polyfill with several improvements and a few breaking changes:
- Far better performance: The polyfill comes with much less runtime costs and more smooth scrolling.
- Better spec compliancy: The polyfill now handles 'Element.prototype.scrollIntoView', works with ShadowDOM, and works with anchor tag navigation.

BREAKING CHANGE: CSS Stylesheets will no longer be parsed. Instead, you must either set inline styles, an attribute with the same name, or set it imperatively. Of course, you can still use the imperative API.